Chatfilter update - NEEDS MORE WORK
I added some better code and some more crap to the old stuff, but it needs lots of work. This cleans up what you guys had before and makes it 10x easier to add more words to the list. It also adds a whitelist to words like "night" don't get mistaken as profane words
Copy what I put in and change it or whatever
inb4 this push is deleted